Sleep Modes and LDR sensor
I made the sleep modes, the switches between them are not perfect yet, but all of the modes are working as intended. The LDR only turns on the LED, if it is dark and only in the active mode. Also I added the 30 secondes without detection of a motion, where we can just add the information of the motion sensor.